more ERC4626 behaviors to a separate behavior file + file border condition (overflow)
