+++ test/token/ERC20/extensions/ERC7540SyncDeposit.test.js
@@ -2,11 +2,8 @@ const { ethers } = require('hardhat');
 const { expect } = require('chai');
 const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers');
 
-const {
-  shouldBehaveLikeERC7540Operator,
-  shouldBehaveLikeERC4626Deposit,
-  shouldBehaveLikeERC7540Redeem,
-} = require('./ERC7540.behavior');
+const { shouldBehaveLikeERC4626Deposit } = require('./ERC4626.behavior');
+const { shouldBehaveLikeERC7540Operator, shouldBehaveLikeERC7540Redeem } = require('./ERC7540.behavior');
 
 const name = 'Vault Shares';
 const symbol = 'vSHR';
